* commit '316af8dd675d777f4bd9512b8af5391c7c89363f':
  Fix crop activity aspect ratio and orientation handling.

+++ src/com/android/gallery3d/filtershow/crop/CropObject.java
@@ -89,7 +89,7 @@ public class CropObject {
         clearSelectState();
     }
 
-    public boolean setInnerAspectRatio(int width, int height) {
+    public boolean setInnerAspectRatio(float width, float height) {
         if (width <= 0 || height <= 0) {
             throw new IllegalArgumentException("Width and Height must be greater than zero");
         }
